add avatar deletion check

+++ src/main/java/one/whr/simple/service/AvatarService.java
@@ -1,8 +1,10 @@
 package one.whr.simple.service;
 
 import one.whr.simple.entity.Avatar;
+import one.whr.simple.entity.User;
 import one.whr.simple.exceptions.AvatarNotFoundException;
 import one.whr.simple.repository.AvatarRepository;
+import one.whr.simple.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,9 +15,12 @@ public class AvatarService {
 
     AvatarRepository avatarRepository;
 
+    UserRepository userRepository;
+
     @Autowired
-    public AvatarService(AvatarRepository avatarRepository) {
+    public AvatarService(AvatarRepository avatarRepository, UserRepository userRepository) {
         this.avatarRepository = avatarRepository;
+        this.userRepository = userRepository;
     }
 
     private volatile Avatar defaultAvatar;
@@ -39,7 +44,19 @@ public class AvatarService {
         return avatarRepository.findAll();
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws AvatarNotFoundException {
+        if (id == 1L) {
+            return;
+        }
+
+        Avatar avatar = avatarRepository.findById(id).orElseThrow(() -> new AvatarNotFoundException("Cannot find avatar to delete"));
+        List<User> users = avatar.getUsers();
+        for (User user : users) {
+            user.setAvatar(getDefaultAvatar());
+        }
+
+        userRepository.saveAll(users);
+
         avatarRepository.deleteById(id);
     }
